fix stacked series when values are below 1

+++ src/modules/Range.js
@@ -107,9 +107,9 @@ class Range {
           if (gl.series[i][j] !== null && Utils.isNumber(gl.series[i][j])) {
             if (gl.series[i][j] > 0) {
               // 0.0001 fixes #185 when values are very small
-              poss = poss + parseInt(gl.series[i][j]) + 0.0001
+              poss = poss + parseFloat(gl.series[i][j]) + 0.0001
             } else {
-              negs = negs + parseInt(gl.series[i][j])
+              negs = negs + parseFloat(gl.series[i][j])
             }
           }
